requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database Drivers & ORM
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0              # PostgreSQL async driver
aiomysql>=0.2.0              # MySQL async driver

# AI/LLM Integration
httpx>=0.27.0  # For Ollama API calls (Local, FREE!)
# openai>=1.0.0  # OpenAI GPT models (Optional, requires billing)
# google-generativeai>=0.8.0  # Deprecated

# Security
python-multipart>=0.0.12

# SQL Parsing & Validation
sqlparse>=0.5.0              # SQL parsing and formatting

# Caching
cachetools>=5.5.0            # In-memory caching

# Utilities
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=24.4.0

# Testing (Optional - uncomment if needed)
# pytest>=8.3.0
# pytest-asyncio>=0.24.0
# httpx>=0.27.0
